+++ KalturaClient/src/main/java/com/kaltura/client/types/AccessControlScope.java
@@ -78,7 +78,7 @@ public class AccessControlScope extends ObjectBase {
 	 * Unix timestamp (In seconds) to be used to test entry scheduling, keep null to
 	  use now.
 	 */
-	private Integer time;
+	private Long time;
 	/**
 	 * Indicates what contexts should be tested. No contexts means any context.
 	 */
@@ -137,10 +137,10 @@ public class AccessControlScope extends ObjectBase {
 	}
 
 	// time:
-	public Integer getTime(){
+	public Long getTime(){
 		return this.time;
 	}
-	public void setTime(Integer time){
+	public void setTime(Long time){
 		this.time = time;
 	}
 
@@ -179,7 +179,7 @@ public class AccessControlScope extends ObjectBase {
 		ip = GsonParser.parseString(jsonObject.get("ip"));
 		ks = GsonParser.parseString(jsonObject.get("ks"));
 		userAgent = GsonParser.parseString(jsonObject.get("userAgent"));
-		time = GsonParser.parseInt(jsonObject.get("time"));
+		time = GsonParser.parseLong(jsonObject.get("time"));
 		contexts = GsonParser.parseArray(jsonObject.getAsJsonArray("contexts"), AccessControlContextTypeHolder.class);
 		hashes = GsonParser.parseArray(jsonObject.getAsJsonArray("hashes"), KeyValue.class);
 
@@ -239,7 +239,7 @@ public class AccessControlScope extends ObjectBase {
         this.ip = in.readString();
         this.ks = in.readString();
         this.userAgent = in.readString();
-        this.time = (Integer)in.readValue(Integer.class.getClassLoader());
+        this.time = (Long)in.readValue(Long.class.getClassLoader());
         int contextsSize = in.readInt();
         if( contextsSize > -1) {
             this.contexts = new ArrayList<>();
