+++ KalturaClient/src/main/java/com/kaltura/client/types/AssetParams.java
@@ -85,7 +85,7 @@ public class AssetParams extends ObjectBase {
 	/**
 	 * Creation date as Unix timestamp (In seconds)
 	 */
-	private Integer createdAt;
+	private Long createdAt;
 	/**
 	 * True if those Flavor Params are part of system defaults
 	 */
@@ -171,7 +171,7 @@ public class AssetParams extends ObjectBase {
 	}
 
 	// createdAt:
-	public Integer getCreatedAt(){
+	public Long getCreatedAt(){
 		return this.createdAt;
 	}
 	// isSystemDefault:
@@ -262,7 +262,7 @@ public class AssetParams extends ObjectBase {
 		name = GsonParser.parseString(jsonObject.get("name"));
 		systemName = GsonParser.parseString(jsonObject.get("systemName"));
 		description = GsonParser.parseString(jsonObject.get("description"));
-		createdAt = GsonParser.parseInt(jsonObject.get("createdAt"));
+		createdAt = GsonParser.parseLong(jsonObject.get("createdAt"));
 		isSystemDefault = GsonParser.parseBoolean(jsonObject.get("isSystemDefault"));
 		tags = GsonParser.parseString(jsonObject.get("tags"));
 		requiredPermissions = GsonParser.parseArray(jsonObject.getAsJsonArray("requiredPermissions"), StringHolder.class);
@@ -332,7 +332,7 @@ public class AssetParams extends ObjectBase {
         this.name = in.readString();
         this.systemName = in.readString();
         this.description = in.readString();
-        this.createdAt = (Integer)in.readValue(Integer.class.getClassLoader());
+        this.createdAt = (Long)in.readValue(Long.class.getClassLoader());
         this.isSystemDefault = (Boolean)in.readValue(Boolean.class.getClassLoader());
         this.tags = in.readString();
         int requiredPermissionsSize = in.readInt();
