+++ KalturaClient/src/main/java/com/kaltura/client/types/AuditTrail.java
@@ -78,7 +78,7 @@ public class AuditTrail extends ObjectBase {
 	}
 
 	private Integer id;
-	private Integer createdAt;
+	private Long createdAt;
 	/**
 	 * Indicates when the data was parsed
 	 */
@@ -113,7 +113,7 @@ public class AuditTrail extends ObjectBase {
 		return this.id;
 	}
 	// createdAt:
-	public Integer getCreatedAt(){
+	public Long getCreatedAt(){
 		return this.createdAt;
 	}
 	// parsedAt:
@@ -292,7 +292,7 @@ public class AuditTrail extends ObjectBase {
 
 		// set members values:
 		id = GsonParser.parseInt(jsonObject.get("id"));
-		createdAt = GsonParser.parseInt(jsonObject.get("createdAt"));
+		createdAt = GsonParser.parseLong(jsonObject.get("createdAt"));
 		parsedAt = GsonParser.parseInt(jsonObject.get("parsedAt"));
 		status = AuditTrailStatus.get(GsonParser.parseInt(jsonObject.get("status")));
 		auditObjectType = AuditTrailObjectType.get(GsonParser.parseString(jsonObject.get("auditObjectType")));
@@ -379,7 +379,7 @@ public class AuditTrail extends ObjectBase {
     public AuditTrail(Parcel in) {
         super(in);
         this.id = (Integer)in.readValue(Integer.class.getClassLoader());
-        this.createdAt = (Integer)in.readValue(Integer.class.getClassLoader());
+        this.createdAt = (Long)in.readValue(Long.class.getClassLoader());
         this.parsedAt = (Integer)in.readValue(Integer.class.getClassLoader());
         int tmpStatus = in.readInt();
         this.status = tmpStatus == -1 ? null : AuditTrailStatus.values()[tmpStatus];
