+++ KalturaClient/src/main/java/com/kaltura/client/types/Beacon.java
@@ -68,7 +68,7 @@ public class Beacon extends ObjectBase {
 	/**
 	 * Beacon update date as Unix timestamp (In seconds)
 	 */
-	private Integer updatedAt;
+	private Long updatedAt;
 	/**
 	 * The object which this beacon belongs to
 	 */
@@ -87,7 +87,7 @@ public class Beacon extends ObjectBase {
 		return this.indexType;
 	}
 	// updatedAt:
-	public Integer getUpdatedAt(){
+	public Long getUpdatedAt(){
 		return this.updatedAt;
 	}
 	// relatedObjectType:
@@ -163,7 +163,7 @@ public class Beacon extends ObjectBase {
 		// set members values:
 		id = GsonParser.parseString(jsonObject.get("id"));
 		indexType = GsonParser.parseString(jsonObject.get("indexType"));
-		updatedAt = GsonParser.parseInt(jsonObject.get("updatedAt"));
+		updatedAt = GsonParser.parseLong(jsonObject.get("updatedAt"));
 		relatedObjectType = BeaconObjectTypes.get(GsonParser.parseString(jsonObject.get("relatedObjectType")));
 		eventType = GsonParser.parseString(jsonObject.get("eventType"));
 		objectId = GsonParser.parseString(jsonObject.get("objectId"));
@@ -213,7 +213,7 @@ public class Beacon extends ObjectBase {
         super(in);
         this.id = in.readString();
         this.indexType = in.readString();
-        this.updatedAt = (Integer)in.readValue(Integer.class.getClassLoader());
+        this.updatedAt = (Long)in.readValue(Long.class.getClassLoader());
         int tmpRelatedObjectType = in.readInt();
         this.relatedObjectType = tmpRelatedObjectType == -1 ? null : BeaconObjectTypes.values()[tmpRelatedObjectType];
         this.eventType = in.readString();
